Added UI to MainWindow and Socket Connections
I've done some basic UI stuff and have added the base for socket based connections, as well as options for the user to use themselves as a server. I used this guy's example: https://github.com/AbleOpus/NetworkingSamples

Though I changed a bit of the code to, of course, make it work with my UI, make it work more asynchronously, and how to handle multiple sockets for the basics of a chat app.

Later on the TODO I need to figure out a way to do images. No idea how that will work, but I might have an idea.