Beginning to add hotkeys
It seems to work fine for the most part, but now I've encountered a bug in the networking portion where the other client that is connected will not receive anything sent.
It seems that any socket.Send does not work and causes an InvalidOperationException, but I'll dig further into this later. Otherwise, hotkeys seem to be working!