Fixed bug with sockets and chat
Turns out the problem was the fact that I forgot I was performing Receive in another thread, and because of that I was deadlocking (when I send, I receive at the same time (for a test)). Thus it makes more sense to update first then send, and use delegates to invoke the update into the main UI thread.

Whoops.