Hotkeys are now added completely
Saving and loading and initializing defaults are completed. Hotkeys are now displayed when the Settings bar pops up
Fixed some bugs regarding hotkeys along the way, as well as making it possible for users to type with their hotkey being used (it's a bad solution, but one that works for now).