50% completion
Added some change to the front end to add immediate testing. Maybe should branched out for that but too late now

This push will contain the addition to T2SUser to help keep track of all connected users for the front-end.
The server now has instruction (TO BE IMPLEMENTED) to interact with updates the same way as the client